Added skips>0 to the if-else chain
handle_datagram had an if(skips>0) statement defining the skip behavior, followed by a long chain of if-else statements defining the non-skip behaviors. The skip behavior included a return statement, but it was the same one that followed all the non-skip behaviors. For the sake of consistency, I attached the first non-skip behavior as an "else if" at the end of "if(skips>0)  { ... }" and removed the duplicate return statement.